Fix bugs caused by the libraries bump
As a side effect, this also fixed the outstanding bug in transaction
management that was happening in the previous state of the project.
Seems it had to do with Neo4j not being able to join the thread bound
transaction and also not being able to create a temporary tx. Surely,
something I fucked up unknowingly.